clean(ZMSKVR-1124): rename package adminstration to administration
