+++ zmsautomation/src/test/java/zms/ataf/ui/pages/admin/administration/AuthoritiesAndLocationsPage.java
@@ -1,4 +1,4 @@
-package zms.ataf.ui.pages.admin.adminstration;
+package zms.ataf.ui.pages.admin.administration;
 
 import java.time.Duration;
 import java.time.LocalDate;
